services/api/requirements.txt
fastapi
uvicorn
kafka-python
elasticsearch
httpx
transformers
protobuf
tiktoken
sentencepiece